requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
lxml==4.9.2
python-telegram-bot==20.0b0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.3.2.post1
urllib3==1.26.13
